pyproject.toml
[tool.poetry]
name = "django-section"
version = "0.1.0"
description = ""
authors = ["Fedorov Semen <FedorovSemen28@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
django-environ = "^0.12.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.0"
dependency-injector = "^4.48.1"
tqdm = "^4.67.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
